--- Charater/Character.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Project1", "Project1\Project1.vcxproj", "{9187B568-694A-450A-8CB7-C75F6033A777}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Debug|x64.ActiveCfg = Debug|x64
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Debug|x64.Build.0 = Debug|x64
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Debug|x86.ActiveCfg = Debug|Win32
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Debug|x86.Build.0 = Debug|Win32
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Release|x64.ActiveCfg = Release|x64
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Release|x64.Build.0 = Release|x64
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Release|x86.ActiveCfg = Release|Win32
-		{9187B568-694A-450A-8CB7-C75F6033A777}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
